// src/app/core/models/FormType.ts
export enum FormType {
  CREATE_SCHEDULE = 'CREATE_SCHEDULE',
  DELETE_SCHEDULE = 'DELETE_SCHEDULE',
  CREATE_GINASIO = 'CREATE_GINASIO',
  DELETE_GINASIO = 'DELETE_GINASIO',
  CREATE_RESTRICAO = 'CREATE_RESTRICAO',
  DELETE_RESTRICAO = 'DELETE_RESTRICAO'
}

export interface ModalConfig {
  type: FormType;
  title: string;
  isOpen: boolean;
}